[FIR] Track local variable assigned types for smartcast stability
When a local variable is captured by a lambda without a contract, future
assignments can impact the smartcast stability for that variable within
the lambda. Because future assignments can appear in both resolved and
unresolved code, track each assignment as encountered to determine if
the smartcast can be considered stable if all future assignments are
resolved and do not revert variable to a super type.

^KT-58191 Fixed