[FIR] Properties defined in a do-while may not always be initialized
Local properties defined within the body of a do-while loop can be used
in the condition of the loop. However, use of a continue may mean the
property isn't always initialized, even if it is initialized when it is
defined. So while a local property may be within scope and has an
initializer, this doesn't always mean that the property is initialized.
Therefore, presence of an initializer should not be used to skip
properties when checking property access via the control-flow graph.

^KT-64872 Fixed